Update Search Bar
I created a search bar component as well as an error message component. Now the user can find a specific recipe based in title search, also if the user inputs an invalid search term, they will be redirected to a page that informs them that the recipe in is not available.